Normalize internal and resource javaagent packages
