+++ instrumentation/internal/internal-application-logger/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/internal/application/logger/ApplicationLoggingInstrumentationModule.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.internal.logging;
+package io.opentelemetry.javaagent.instrumentation.internal.application.logger;
 
 import static java.util.Arrays.asList;
