+++ instrumentation/spring/spring-boot-resources/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/boot/resources/SystemHelper.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.spring.resources;
+package io.opentelemetry.javaagent.instrumentation.spring.boot.resources;
 
 import static java.util.logging.Level.FINER;
